Make court cases by visible forces query reusable
